fix(supabase): store commitment as string
postgresql bigint type can't hold the JS bigint types.
bigint is harder to deal with when making api calls to the db which
needs serialized json data
