+++ supabase/package.json
@@ -4,7 +4,7 @@
     "restart": "supabase stop && supabase start",
     "start": "supabase start",
     "stop": "supabase stop",
-    "types": "supabase gen types --lang=typescript --local > ../apps/server/src/supabase/supabase-generated.types.ts"
+    "types": "supabase gen types --lang=typescript --local > ../apps/server/src/supabase/supabase.types.ts"
   },
   "devDependencies": {
     "mkdirp": "^3.0.1",
